Rename MockedPermissionUtils to FakePermissionUtils
